create Input for search movie on React
